requirements.txt
python-telegram-bot[webhooks]>=21.0
cryptography>=43.0
httpx>=0.27
pyyaml>=6.0
python-frontmatter>=1.1
python-dotenv>=1.0
